Renamed button style file
